LRSD-7193 Refactor permissions
